requirements.txt
# AT Protocol HMA Integration Dependencies

# HMA Dependencies
threatexchange>=1.0.0
hasher-matcher-actioner>=2.0.0

# AT Protocol Dependencies
atproto>=0.0.1

# Web Framework
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
psycopg2-binary>=2.9.6

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pillow>=10.0.0
pyyaml>=6.0.0
loguru>=0.7.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.1 